L1-support-tools/06-bulk-action-runner/macos/pyproject.toml
[project]
name = "bulk-action-runner"
version = "0.1.0"
description = "CSV-powered bulk operations tool for L1 support"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "msgraph-sdk>=1.0.0",
    "azure-identity>=1.14.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
bulk-run = "bulk.__main__:cli"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["bulk*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
